Add files via upload
This is my own invention of blackjack code. i did my best to understand the Black Jack system. i did not use any function or that much class because i was still groping this kinds of topics. Please tell what is your opinion regarding my source code. I personally believe the harder the criticism the easier the lesson to learn.